v1.0.3

+++ common/user.go
@@ -3,5 +3,5 @@ package common
 type LoginUser struct {
 	Id   int
 	Name string
-	Role byte
+	Role int
 }
